provider/gce: Duplicate Operation{Waiter,Error}
Type safety means that `HttpsHealthCheck`, which uses the `compute/v0.beta`
library instead of `compute/v1`, can't re-use `OperationWaiter` and
`OperationError` because all of the objects passed in are considered to be
of a different type. So we have to duplicate them to accept the correct
types.

I did attempt to use empty interfaces and type assertions but I still needed
to duplicate the entire `switch` statement in `RefreshFunc()` and it looked
even more complicated and inefficient than this.

Although somewhat unconventional I have duplicated the structs and methods
one-by-one rather than adding everything to the end of the file. This is in
the hope that it makes it more obvious that you need to update both if you
make a change.

If anyone finds a nicer way to do this in the future then I would be more
than happy for it to be refactored.