provider/google: TargetPool use self_link for HCs
Change `google_compute_target_pool` to use the `self_link` attribute instead
of `name` when referencing `google_compute_http_health_check` resources.

This is to aide the addition of `google_compute_https_health_check`. HTTP
and HTTPS health checks can have the same name, so it's no longer safe or
possible for us to lookup the resource dependencies by name.

NB: This is a breaking change.